+++ src/WriterModule/ns10/ns10_Writer.cpp
@@ -19,53 +19,27 @@ namespace ns10 {
 static WriterModule::Registry::Registrar<ns10_Writer> RegisterWriter("ns10",
                                                                      "ns10");
 
-void ns10_Writer::parse_config(std::string const &ConfigurationStream) {
-  auto Config = nlohmann::json::parse(ConfigurationStream);
-  try {
-    CueInterval = Config["cue_interval"].get<uint64_t>();
-  } catch (std::exception const &Exception) {
-    Logger->warn("Unable to extract cue interval from JSON.");
-  }
-
-  Logger->info("Using a cue interval of {}.", CueInterval);
-
-  auto JsonChunkSize = Config["chunk_size"];
-  if (JsonChunkSize.is_array()) {
-    ChunkSize = Config["chunk_size"].get<hdf5::Dimensions>();
-  } else if (JsonChunkSize.is_number_integer()) {
-    ChunkSize = hdf5::Dimensions{JsonChunkSize.get<hsize_t>()};
-  } else {
-    Logger->warn("Unable to extract chunk size, using the existing value");
-  }
-  try {
-    Sourcename = Config["source"];
-  } catch (std::exception const &Exception) {
-    Logger->error("Key \"source\" is not specified in JSON command");
-  }
-}
-
 using FileWriterBase = WriterModule::Base;
 
 WriterModule::InitResult ns10_Writer::init_hdf(hdf5::node::Group &HDFGroup) {
-  const int DefaultChunkSize = ChunkSize.at(0);
   try {
     auto &CurrentGroup = HDFGroup;
     NeXusDataset::DoubleValue(      // NOLINT(bugprone-unused-raii)
         CurrentGroup,               // NOLINT(bugprone-unused-raii)
         NeXusDataset::Mode::Create, // NOLINT(bugprone-unused-raii)
-        DefaultChunkSize);          // NOLINT(bugprone-unused-raii)
+        ChunkSize);                 // NOLINT(bugprone-unused-raii)
     NeXusDataset::Time(             // NOLINT(bugprone-unused-raii)
         CurrentGroup,               // NOLINT(bugprone-unused-raii)
         NeXusDataset::Mode::Create, // NOLINT(bugprone-unused-raii)
-        DefaultChunkSize);          // NOLINT(bugprone-unused-raii)
+        ChunkSize);                 // NOLINT(bugprone-unused-raii)
     NeXusDataset::CueIndex(         // NOLINT(bugprone-unused-raii)
         CurrentGroup,               // NOLINT(bugprone-unused-raii)
         NeXusDataset::Mode::Create, // NOLINT(bugprone-unused-raii)
-        DefaultChunkSize);          // NOLINT(bugprone-unused-raii)
+        ChunkSize);                 // NOLINT(bugprone-unused-raii)
     NeXusDataset::CueTimestampZero( // NOLINT(bugprone-unused-raii)
         CurrentGroup,               // NOLINT(bugprone-unused-raii)
         NeXusDataset::Mode::Create, // NOLINT(bugprone-unused-raii)
-        DefaultChunkSize);          // NOLINT(bugprone-unused-raii)
+        ChunkSize);                 // NOLINT(bugprone-unused-raii)
   } catch (std::exception &E) {
     Logger->error("Unable to initialise areaDetector data tree in "
                   "HDF file with error message: \"{}\"",
@@ -99,11 +73,6 @@ void ns10_Writer::write(const FileWriter::FlatbufferMessage &Message) {
   auto Value = Entry->value();
   if (!Source || !Value) {
     throw std::runtime_error("Invalid Flatbuffer content.");
-  }
-
-  if (Source->str() != Sourcename) {
-    Logger->warn("Invalid source name: {}", Source->str());
-    return;
   }
 
   try {