+++ src/tests/WriterModule/NDAr_WriterTests.cpp
@@ -222,7 +222,7 @@ TEST_F(AreaDetectorWriter, WriterDefaultValuesTest) {
   EXPECT_EQ(Dataspace.current_dimensions(), (hdf5::Dimensions{0, 1, 1}));
   auto CreationProperties = Temp.Values->creation_list();
   auto ChunkDims = CreationProperties.chunk();
-  EXPECT_EQ(ChunkDims, (hdf5::Dimensions{64, 1, 1}));
+  EXPECT_EQ(ChunkDims, (hdf5::Dimensions{1048576, 1, 1}));
 }
 
 TEST_F(AreaDetectorWriter, WriterWriteTest) {
@@ -365,9 +365,9 @@ TEST_F(AreaDetectorWriter, ConfigArraySizeTest) {
     "array_size": [5,5,5]
   })"");
   ADWriterStandIn Writer;
-  EXPECT_EQ(Writer.ArrayShape, (hdf5::Dimensions{1, 1}));
+  EXPECT_EQ(Writer.ArrayShape.getValue(), (hdf5::Dimensions{1, 1}));
   Writer.parse_config(JsonConfig.dump());
-  EXPECT_EQ(Writer.ArrayShape, (hdf5::Dimensions{5, 5, 5}));
+  EXPECT_EQ(Writer.ArrayShape.getValue(), (hdf5::Dimensions{5, 5, 5}));
 }
 
 TEST_F(AreaDetectorWriter, ConfigArraySizeFailureTest) {
@@ -375,19 +375,19 @@ TEST_F(AreaDetectorWriter, ConfigArraySizeFailureTest) {
     "array_size": "hello"
   })"");
   ADWriterStandIn Writer;
-  EXPECT_EQ(Writer.ArrayShape, (hdf5::Dimensions{1, 1}));
+  EXPECT_EQ(Writer.ArrayShape.getValue(), (hdf5::Dimensions{1, 1}));
   Writer.parse_config(JsonConfig.dump());
-  EXPECT_EQ(Writer.ArrayShape, (hdf5::Dimensions{1, 1}));
+  EXPECT_EQ(Writer.ArrayShape.getValue(), (hdf5::Dimensions{1, 1}));
 }
 
 TEST_F(AreaDetectorWriter, ConfigChunkSizeTestAlt) {
   auto JsonConfig = nlohmann::json::parse(R""({
-    "chunk_size": 1024
+    "chunk_size": [2048]
   })"");
   ADWriterStandIn Writer;
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{64}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{1048576}));
   Writer.parse_config(JsonConfig.dump());
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{1024}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{2048}));
 }
 
 TEST_F(AreaDetectorWriter, ConfigChunkSizeTest) {
@@ -395,9 +395,9 @@ TEST_F(AreaDetectorWriter, ConfigChunkSizeTest) {
     "chunk_size": [5,5,5,5]
   })"");
   ADWriterStandIn Writer;
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{64}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{1048576}));
   Writer.parse_config(JsonConfig.dump());
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{5, 5, 5, 5}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{5, 5, 5, 5}));
 }
 
 TEST_F(AreaDetectorWriter, ConfigChunkSizeFailureTest) {
@@ -405,9 +405,9 @@ TEST_F(AreaDetectorWriter, ConfigChunkSizeFailureTest) {
     "chunk_size": "hello"
   })"");
   ADWriterStandIn Writer;
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{64}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{1048576}));
   Writer.parse_config(JsonConfig.dump());
-  EXPECT_EQ(Writer.ChunkSize, (hdf5::Dimensions{64}));
+  EXPECT_EQ(Writer.ChunkSize.getValue(), (hdf5::Dimensions{1048576}));
 }
 
 // Note, you must feed it 1000 elements in total