+++ changes.md
@@ -21,3 +21,7 @@ approx 5%, for details of test see PR ([#558](https://github.com/ess-dmsc/kafka-
 - The *NX_class* value configured for a writer module will always override that of the HDF group *NX_class* value. If configured, a writer module attribute (e.g. *NX_class*) will override both.
 - The Kafka topic and flatbuffer source names will now automatically be written as (HDF) attributes when instantiating a new writer module.
 - Minor fix to CMake code for working around differences in how file-name leading lower and upper case letters are handled on different systems.
+- The writer module configuration names/keys have been unified.
+- Better documentation of writer module configuration options.
+- The application will now print an error message if there is a configuration that is not used (due to e.g. a typo).
+- The error reporting and handling of writer module configurations have overall been greatly improved.